requirements.txt
lxml>=4.4
python-dateutil==2.8
requests==2.31.0
numpy
tqdm
pandas

